+++ ts-sdk/cli/src/commands/set_default_clp_to_olp_reward_ratio.ts
@@ -1,27 +1,28 @@
 import {
   fetchMaybeFusionPoolsConfig,
   getFusionPoolsConfigAddress,
-  getSetDefaultOrderProtocolFeeRateInstruction,
+  getSetDefaultClpToOlpRewardRatioInstruction,
 } from "@crypticdot/fusionamm-client";
 import { sendTransaction } from "@crypticdot/fusionamm-tx-sender";
-import BaseCommand, { addressArg } from "../base";
+import BaseCommand from "../base";
 import { rpc, signer } from "../rpc";
 import { Args } from "@oclif/core";
 
-export default class SetDefaultOrderProtocolFeeRate extends BaseCommand {
+export default class SetDefaultClpToOlpRewardRatio extends BaseCommand {
   static override args = {
-    defaultOrderProtocolFeeRate: Args.integer({
-      description: "Limit order fee rate stored as basis points. The maximum value of 10000 equals to 100%",
+    defaultClpToOlpRewardRatio: Args.integer({
+      description: "The default CLP to OLP reward ratio. The maximum value of 10000 equals to 100%",
       required: true,
       min: 0,
       max: 10000,
     }),
   };
-  static override description = "Set the default limit order protocol fee rate";
+  static override description =
+    "Set the default reward ratio of concentrated liquidity providers to limit orders' liquidity providers";
   static override examples = ["<%= config.bin %> <%= command.id %> address 5000"];
 
   public async run() {
-    const { args } = await this.parse(SetDefaultOrderProtocolFeeRate);
+    const { args } = await this.parse(SetDefaultClpToOlpRewardRatio);
 
     const fusionPoolsConfig = (await getFusionPoolsConfigAddress())[0];
 
@@ -32,14 +33,14 @@ export default class SetDefaultOrderProtocolFeeRate extends BaseCommand {
       throw new Error("FusionAMM config doesn't exists at address " + fusionPoolsConfig);
     }
 
-    const ix = getSetDefaultOrderProtocolFeeRateInstruction({
+    const ix = getSetDefaultClpToOlpRewardRatioInstruction({
       fusionPoolsConfig,
       feeAuthority: signer,
-      defaultOrderProtocolFeeRate: args.defaultOrderProtocolFeeRate,
+      defaultClpToOlpRewardRatio: args.defaultClpToOlpRewardRatio,
     });
 
     console.log("");
-    if (config.data.defaultOrderProtocolFeeRate != args.defaultOrderProtocolFeeRate) {
+    if (config.data.defaultClpToOlpRewardRatio != args.defaultClpToOlpRewardRatio) {
       console.log("Sending a transaction...");
       const signature = await sendTransaction(rpc, [ix], signer);
       console.log("Transaction landed:", signature);